\name{CMP-package}
\alias{CMP-package}
\alias{CMP}
\docType{package}
\title{
Conway-Maxwell-Poisson distribution
}
\description{
\packageDescription{CMP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CMP}
\packageIndices{CMP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CMP}

Maintainer: \packageMaintainer{CMP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
